Move Python files to src and break up into smaller files
