+++ src/ADBSync/FileSystems/__init__.py
@@ -1,0 +1,3 @@
+#!/usr/bin/env python3
+
+"""Filesystems classes go in this module"""
