+++ src/ADBSync/__init__.py
@@ -2,357 +2,19 @@
 
 """Better version of adb-sync for Python3"""
 
-__version__ = "1.1.3"
+__version__ = "1.1.4"
 
+from typing import List, Tuple, Union
 import logging
-from typing import Iterator, List, Tuple, Union, Iterable
 import os
-import stat
-import re
-import subprocess
-import datetime
 import fnmatch
 
-from SAOLogging import getParser, setupRootLogger, criticalLogExit, logTree
+from .argparsing import getArgs
+from .SAOIO import tryLoading
+from .SAOLogging import criticalLogExit, logTree, setupRootLogger
 
-class FileSystem():
-    RE_ADB_FILE_PUSHED = re.compile("^.*: 1 file pushed, 0 skipped\\..*$")
-    RE_ADB_FILE_PULLED = re.compile("^.*: 1 file pulled, 0 skipped\\..*$")
-
-    def __init__(self, adb_arguments: List[str]) -> None:
-        self.adb_arguments = adb_arguments
-
-    def adbShell(self, commands: List[str]) -> Iterator[str]:
-        with subprocess.Popen(
-            self.adb_arguments + ["shell"] + commands,
-            stdout = subprocess.PIPE, stderr = subprocess.STDOUT
-        ) as proc:
-            while adbLine := proc.stdout.readline().decode().rstrip("\r\n"):
-                yield adbLine
-
-    def unlink(self, path: str) -> None:
-        raise NotImplementedError
-
-    def rmdir(self, path: str) -> None:
-        raise NotImplementedError
-
-    def makedirs(self, path: str) -> None:
-        raise NotImplementedError
-
-    def realPath(self, path: str) -> str:
-        raise NotImplementedError
-
-    def lstat(self, path: str) -> os.stat_result:
-        raise NotImplementedError
-
-    def lstat_inDir(self, path: str) -> Iterable[Tuple[str, os.stat_result]]:
-        raise NotImplementedError
-
-    def utime(self, path: str, times: Tuple[int, int]) -> None:
-        raise NotImplementedError
-
-    @staticmethod
-    def joinPaths(base: str, leaf: str) -> str:
-        raise NotImplementedError
-
-    @staticmethod
-    def normPath(path: str) -> str:
-        raise NotImplementedError
-
-    def pushFileHere(self, source: str, destination: str) -> None:
-        raise NotImplementedError
-
-    def _getFilesTree(self, tree_root: str, tree_root_stat: os.stat_result, followLinks: bool = False):
-        # the reason to have two functions instead of one purely recursive one is to use self.lstat_inDir ie ls
-        # which is much faster than individually stat-ing each file. Hence we have getFilesTree's special first lstat
-        if stat.S_ISLNK(tree_root_stat.st_mode):
-            if not followLinks:
-                logging.warning("Ignoring symlink '{}'".format(tree_root))
-                return None
-            logging.debug("Following symlink '{}'".format(tree_root))
-            try:
-                tree_root_realPath = self.realPath(tree_root)
-                tree_root_stat_realPath = self.lstat(tree_root_realPath)
-            except FileNotFoundError:
-                logging.error("Skipping dead symlink '{}'".format(tree_root))
-                return None
-            except NotADirectoryError:
-                logging.error("Skipping not-a-directory symlink '{}'".format(tree_root))
-                return None
-            return self._getFilesTree(tree_root_realPath, tree_root_stat_realPath, followLinks = followLinks)
-        elif stat.S_ISDIR(tree_root_stat.st_mode):
-            tree = {".": (60 * (int(tree_root_stat.st_atime) // 60), 60 * (int(tree_root_stat.st_mtime) // 60))}
-            for filename, statObject_child, in self.lstat_inDir(tree_root):
-                if filename in [".", ".."]:
-                    continue
-                tree[filename] = self._getFilesTree(
-                    self.joinPaths(tree_root, filename),
-                    statObject_child,
-                    followLinks = followLinks)
-            return tree
-        elif stat.S_ISREG(tree_root_stat.st_mode):
-            return (60 * (int(tree_root_stat.st_atime) // 60), 60 * (int(tree_root_stat.st_mtime) // 60))
-        else:
-            raise NotImplementedError
-
-    def getFilesTree(self, tree_root: str, followLinks: bool = False):
-        statObject = self.lstat(tree_root)
-        return self._getFilesTree(tree_root, statObject, followLinks = followLinks)
-
-    def removeTree(self, tree_root: str, tree: Union[Tuple[int, int], dict], dry_run: bool = True) -> None:
-        if isinstance(tree, tuple):
-            logging.info("Removing '{}'".format(tree_root))
-            if not dry_run:
-                self.unlink(tree_root)
-        elif isinstance(tree, dict):
-            removeFolder = tree.pop(".", False)
-            for key, value in tree.items():
-                self.removeTree(self.normPath(self.joinPaths(tree_root, key)), value, dry_run = dry_run)
-            if removeFolder:
-                logging.info("Removing folder '{}'".format(tree_root))
-                if not dry_run:
-                    self.rmdir(tree_root)
-        else:
-            raise NotImplementedError
-
-    def pushTreeHere(self,
-        tree_root: str,
-        tree: Union[Tuple[int, int], dict],
-        destination_root: str,
-        pathJoinFunction_source,
-        pathNormFunction_source,
-        pathJoinFunction_destination,
-        pathNormFunction_destination,
-        dry_run: bool = True
-        ) -> None:
-        if isinstance(tree, tuple):
-            logging.info("Copying '{}' to '{}'".format(tree_root, destination_root))
-            if not dry_run:
-                self.pushFileHere(tree_root, destination_root)
-                self.utime(destination_root, tree)
-        elif isinstance(tree, dict):
-            if tree.pop(".", False):
-                logging.info("Making directory '{}'".format(destination_root))
-                if not dry_run:
-                    self.makedirs(destination_root)
-            for key, value in tree.items():
-                self.pushTreeHere(
-                    pathNormFunction_source(pathJoinFunction_source(tree_root, key)),
-                    value,
-                    pathNormFunction_destination(pathJoinFunction_destination(destination_root, key)),
-                    pathJoinFunction_source,
-                    pathNormFunction_source,
-                    pathJoinFunction_destination,
-                    pathNormFunction_destination,
-                    dry_run = dry_run
-                )
-        else:
-            raise NotImplementedError
-
-class LocalFileSystem(FileSystem):
-    def unlink(self, path: str) -> None:
-        os.unlink(path)
-
-    def rmdir(self, path: str) -> None:
-        os.rmdir(path)
-
-    def makedirs(self, path: str) -> None:
-        os.makedirs(path, exist_ok = True)
-
-    def realPath(self, path: str) -> str:
-        return os.path.realpath(path)
-
-    def lstat(self, path: str) -> os.stat_result:
-        return os.lstat(path)
-
-    def lstat_inDir(self, path: str) -> Iterable[Tuple[str, os.stat_result]]:
-        for filename in os.listdir(path):
-            yield filename, self.lstat(self.joinPaths(path, filename))
-
-    def utime(self, path: str, times: Tuple[int, int]) -> None:
-        os.utime(path, times)
-
-    @staticmethod
-    def joinPaths(base: str, leaf: str) -> str:
-        return os.path.join(base, leaf)
-
-    @staticmethod
-    def normPath(path: str) -> str:
-        return os.path.normpath(path)
-
-    def pushFileHere(self, source: str, destination: str) -> None:
-        with subprocess.Popen(
-            self.adb_arguments + ["pull", source, destination],
-            stdout = subprocess.PIPE, stderr = subprocess.STDOUT
-        ) as proc:
-            while adbLine := proc.stdout.readline().decode().rstrip("\r\n"):
-                if not self.RE_ADB_FILE_PULLED.fullmatch(adbLine):
-                    criticalLogExit("Line not captured: '{}'".format(adbLine))
-
-class AndroidFileSystem(FileSystem):
-    RE_TESTCONNECTION_NO_DEVICE = re.compile("^adb\\: no devices/emulators found$")
-    RE_TESTCONNECTION_DAEMON_NOT_RUNNING = re.compile("^\\* daemon not running; starting now at tcp:\\d+$")
-    RE_TESTCONNECTION_DAEMON_STARTED = re.compile("^\\* daemon started successfully$")
-
-    RE_LS_TO_STAT = re.compile(
-        r"""^
-        (?:
-        (?P<S_IFREG> -) |
-        (?P<S_IFBLK> b) |
-        (?P<S_IFCHR> c) |
-        (?P<S_IFDIR> d) |
-        (?P<S_IFLNK> l) |
-        (?P<S_IFIFO> p) |
-        (?P<S_IFSOCK> s))
-        [-r][-w][-xsS]
-        [-r][-w][-xsS]
-        [-r][-w][-xtT] # Mode string
-        [ ]+
-        (?:
-        [0-9]+ # Number of hard links
-        [ ]+
-        )?
-        [^ ]+ # User name/ID
-        [ ]+
-        [^ ]+ # Group name/ID
-        [ ]+
-        (?(S_IFBLK) [^ ]+[ ]+[^ ]+[ ]+) # Device numbers
-        (?(S_IFCHR) [^ ]+[ ]+[^ ]+[ ]+) # Device numbers
-        (?(S_IFDIR) (?P<dirsize>[0-9]+ [ ]+))? # Directory size
-        (?(S_IFREG) (?P<st_size> [0-9]+) [ ]+) # Size
-        (?(S_IFLNK) ([0-9]+) [ ]+) # Link length
-        (?P<st_mtime>
-        [0-9]{4}-[0-9]{2}-[0-9]{2} # Date
-        [ ]
-        [0-9]{2}:[0-9]{2}) # Time
-        [ ]
-        # Don't capture filename for symlinks (ambiguous).
-        (?(S_IFLNK) .* | (?P<filename> .*))
-        $""", re.DOTALL | re.VERBOSE)
-
-    RE_NO_SUCH_FILE = re.compile("^.*: No such file or directory$")
-    RE_LS_NOT_A_DIRECTORY = re.compile("ls: .*: Not a directory$")
-    RE_TOTAL = re.compile("^total \\d+$")
-
-    RE_REALPATH_NO_SUCH_FILE = re.compile("^realpath: .*: No such file or directory$")
-    RE_REALPATH_NOT_A_DIRECTORY = re.compile("^realpath: .*: Not a directory$")
-
-    escapePathReplacements = [
-        [" ", "\\ "],
-        ["'", "\\'"],
-        ["(", "\\("],
-        [")", "\\)"],
-        ["!", "\\!"],
-        ["&", "\\&"]
-    ]
-
-    def unlink(self, path: str) -> None:
-        for line in self.adbShell(["rm", self.escapePath(path)]):
-            if line:
-                criticalLogExit("Line not captured: '{}'".format(line))
-
-    def rmdir(self, path: str) -> None:
-        for line in self.adbShell(["rm -r", self.escapePath(path)]):
-            if line:
-                criticalLogExit("Line not captured: '{}'".format(line))
-
-    def makedirs(self, path: str) -> None:
-        for line in self.adbShell(["mkdir -p", self.escapePath(path)]):
-            if line:
-                criticalLogExit("Line not captured: '{}'".format(line))
-
-    def realPath(self, path: str) -> str:
-        for line in self.adbShell(["realpath", self.escapePath(path)]):
-            if self.RE_REALPATH_NO_SUCH_FILE.fullmatch(line):
-                raise FileNotFoundError
-            elif self.RE_REALPATH_NOT_A_DIRECTORY.fullmatch(line):
-                raise NotADirectoryError
-            else:
-                return line
-
-    def lstat(self, path: str) -> os.stat_result:
-        for line in self.adbShell(["ls -lad", self.escapePath(path)]):
-            return self.lsToStat(line)[1]
-
-    def lstat_inDir(self, path: str) -> Iterable[Tuple[str, os.stat_result]]:
-        for line in self.adbShell(["ls", "-la", self.escapePath(path)]):
-            if self.RE_TOTAL.fullmatch(line):
-                continue
-            else:
-                yield self.lsToStat(line)
-
-    def utime(self, path: str, times: Tuple[int, int]) -> None:
-        atime = datetime.datetime.utcfromtimestamp(times[0]).strftime("%Y%m%d%H%M")
-        mtime = datetime.datetime.utcfromtimestamp(times[1]).strftime("%Y%m%d%H%M")
-        for line in self.adbShell(["touch", "-at", atime, "-mt", mtime, self.escapePath(path)]):
-            criticalLogExit("Line not captured: '{}'".format(line))
-
-    @staticmethod
-    def joinPaths(base: str, leaf: str) -> str:
-        return os.path.join(base, leaf).replace("\\", "/") # for Windows
-
-    @staticmethod
-    def normPath(path: str) -> str:
-        return os.path.normpath(path).replace("\\", "/")
-
-    def pushFileHere(self, source: str, destination: str) -> None:
-        with subprocess.Popen(
-            self.adb_arguments + ["push", source, destination],
-            stdout = subprocess.PIPE, stderr = subprocess.STDOUT
-        ) as proc:
-            while adbLine := proc.stdout.readline().decode().rstrip("\r\n"):
-                if not self.RE_ADB_FILE_PUSHED.fullmatch(adbLine):
-                    criticalLogExit("Line not captured: '{}'".format(adbLine))
-
-    def escapePath(self, path: str) -> str:
-        for replacement in self.escapePathReplacements:
-            path = path.replace(*replacement)
-        return path
-
-    def testConnection(self):
-        for line in self.adbShell([":"]):
-            if self.RE_TESTCONNECTION_DAEMON_NOT_RUNNING.fullmatch(line) or self.RE_TESTCONNECTION_DAEMON_STARTED.fullmatch(line):
-                continue
-            elif self.RE_TESTCONNECTION_NO_DEVICE.fullmatch(line):
-                return False
-        return True
-
-    def lsToStat(self, line: str) -> Tuple[str, os.stat_result]:
-        if self.RE_NO_SUCH_FILE.fullmatch(line):
-            raise FileNotFoundError
-        elif self.RE_LS_NOT_A_DIRECTORY.fullmatch(line):
-            raise NotADirectoryError
-        elif match := self.RE_LS_TO_STAT.fullmatch(line):
-            match_groupdict = match.groupdict()
-            st_mode = stat.S_IRWXU | stat.S_IRGRP | stat.S_IXGRP | stat.S_IROTH | stat.S_IXOTH # 755
-            if match_groupdict['S_IFREG']:
-                st_mode |= stat.S_IFREG
-            if match_groupdict['S_IFBLK']:
-                st_mode |= stat.S_IFBLK
-            if match_groupdict['S_IFCHR']:
-                st_mode |= stat.S_IFCHR
-            if match_groupdict['S_IFDIR']:
-                st_mode |= stat.S_IFDIR
-            if match_groupdict['S_IFIFO']:
-                st_mode |= stat.S_IFIFO
-            if match_groupdict['S_IFLNK']:
-                st_mode |= stat.S_IFLNK
-            if match_groupdict['S_IFSOCK']:
-                st_mode |= stat.S_IFSOCK
-            st_size = None if match_groupdict["st_size"] is None else int(match_groupdict["st_size"])
-            st_mtime = int(datetime.datetime.strptime(match_groupdict["st_mtime"], "%Y-%m-%d %H:%M").timestamp())
-
-            # Fill the rest with dummy values.
-            st_ino = 1
-            st_rdev = 0
-            st_nlink = 1
-            st_uid = -2  # Nobody.
-            st_gid = -2  # Nobody.
-            st_atime = st_ctime = st_mtime
-
-            return match_groupdict["filename"], os.stat_result((st_mode, st_ino, st_rdev, st_nlink, st_uid, st_gid, st_size, st_atime, st_mtime, st_ctime))
-        else:
-            criticalLogExit("Line not captured: '{}'".format(line))
+from .FileSystems.Local import LocalFileSystem
+from .FileSystems.Android import AndroidFileSystem
 
 class FileSyncer():
     @classmethod
@@ -617,84 +279,22 @@ class FileSyncer():
                     returnDict[key] = value_pruned
             return returnDict or None
 
-if __name__ == "__main__":
-    parser = getParser(__doc__, version = __version__)
+def main():
+    args = getArgs(__doc__, __version__)
 
-    parser.add_argument("LOCAL",
-        help = "Local path",
-        action = "store")
-    parser.add_argument("ANDROID",
-        help = "Android path",
-        action = "store")
-    parser.add_argument("-n", "--dry-run",
-        help = "Perform a dry run; do not actually copy and delete etc",
-        action = "store_true",
-        default = False)
-    parser.add_argument("-L", "--copy-links",
-        help = "Follow symlinks and copy their referent file / directory",
-        action = "store_true",
-        default = False)
-    parser.add_argument("--exclude",
-        help = "fnmatch pattern to ignore relative to source (reusable)",
-        action = "append",
-        default = [])
-    parser.add_argument("--exclude-from",
-        help = "Filename of file containing fnmatch patterns to ignore relative to source (reusable)",
-        action = "append",
-        default = [])
-    parser.add_argument("--del",
-        help = "Delete files at the destination that are not in the source",
-        action = "store_true",
-        dest = "delete",
-        default = False)
-    parser.add_argument("--delete-excluded",
-        help = "Delete files at the destination that are excluded",
-        action = "store_true",
-        default = False)
-    parser.add_argument("--pull",
-        help = "Pull ANDROID from Android to LOCAL on the computer instead of the default pushing from computer to Android",
-        action = "store_true",
-        default = False)
-    parser.add_argument("--force",
-        help = "Allows files to overwrite folders and folders to overwrite files. This is false by default to prevent large scale accidents",
-        action = "store_true",
-        default = False)
-
-    parser_group_adb = parser.add_argument_group("ADB arguments",
-        description = "By default ADB works for me without touching any of these, but if you have any specific "
-        "demands then go ahead. See 'adb --help' for a full list of adb flags and options")
-    parser_group_adb.add_argument("--adb-bin",
-        help = "Use the given adb binary. Defaults to 'adb' ie whatever is on path",
-        default = "adb")
-    parser_group_adb.add_argument("--adb-flag",
-        help = "Add a flag to call adb with, eg '--adb-flag d' for adb -d, that is return an error if more than one device is connected",
-        metavar = "ADB_FLAG",
-        action = "append",
-        dest = "adb_flags",
-        default = [])
-    parser_group_adb.add_argument("--adb-option",
-        help = "Add an option to call adb with, eg '--adb-option P 5037' for adb -P 5037, that is use port 5037 for the adb server",
-        metavar = ("OPTION", "VALUE"),
-        nargs = 2,
-        action = "append",
-        dest = "adb_options",
-        default = [])
-
-    args = parser.parse_args()
-
-    setupRootLogger(args.verbose, args.quiet, messagefmt = "[%(levelname)s] %(message)s")
+    setupRootLogger(
+        noColor = args.logging_noColor,
+        verbosityLevel = args.logging_verbosity_verbose,
+        quietnessLevel = args.logging_verbosity_quiet,
+        messagefmt = "[%(levelname)s] %(message)s"
+    )
 
     if args.LOCAL[-1] in ["/", "\\"] or args.ANDROID[-1] in ["/", "\\"]:
         logging.warning("Trailing slashes are ignored (see README.md)")
 
-    for exclude_from_filename in args.exclude_from:
-        try:
-            with open(os.path.expanduser(os.path.normpath(exclude_from_filename)), "r") as f:
-                while line := f.readline():
-                    if line_stripped := line.rstrip("\r\n"):
-                        args.exclude.append(line_stripped)
-        except (FileNotFoundError, PermissionError):
-            criticalLogExit("Could not open exclude-from file '{}'".format(exclude_from_filename))
+    for excludeFrom_filename in args.excludeFrom:
+        excludeFrom_file = tryLoading(os.path.expanduser(os.path.normpath(excludeFrom_filename)))
+        args.exclude.extend([line for line in excludeFrom_file.splitlines() if line])
 
     adb_arguments = [args.adb_bin] + ["-{}".format(arg) for arg in args.adb_flags]
     for option, value in args.adb_options:
@@ -723,7 +323,7 @@ if __name__ == "__main__":
         criticalLogExit("No device detected")
 
     try:
-        filesTree_source = fs_source.getFilesTree(path_source, followLinks = args.copy_links)
+        filesTree_source = fs_source.getFilesTree(path_source, followLinks = args.copyLinks)
     except FileNotFoundError:
         criticalLogExit("Source '{}' not found".format(path_source))
     except NotADirectoryError:
@@ -732,7 +332,7 @@ if __name__ == "__main__":
         criticalLogExit("Permission error stat-ing '{}'".format(path_source))
 
     try:
-        filesTree_destination = fs_destination.getFilesTree(path_destination, followLinks = args.copy_links)
+        filesTree_destination = fs_destination.getFilesTree(path_destination, followLinks = args.copyLinks)
     except FileNotFoundError:
         filesTree_destination = None
     except NotADirectoryError:
@@ -770,7 +370,7 @@ if __name__ == "__main__":
         excludePatterns,
         fs_source.joinPaths,
         fs_destination.joinPaths,
-        folderFileOverwriteError = not args.dry_run and not args.force
+        folderFileOverwriteError = not args.dryRun and not args.force
     )
 
     tree_delete = FileSyncer.pruneTree(tree_delete)
@@ -824,35 +424,35 @@ if __name__ == "__main__":
 
     if tree_delete is not None:
         logging.info("Deleting delete tree")
-        fs_destination.removeTree(path_destination, tree_delete, dry_run = args.dry_run)
+        fs_destination.removeTree(path_destination, tree_delete, dryRun = args.dryRun)
     else:
         logging.info("Empty delete tree")
     logging.info("")
 
-    if args.delete_excluded and args.delete:
+    if args.deleteExcluded and args.delete:
         if tree_excluded_destination is not None:
             logging.info("Deleting destination excluded tree")
-            fs_destination.removeTree(path_destination, tree_excluded_destination, dry_run = args.dry_run)
+            fs_destination.removeTree(path_destination, tree_excluded_destination, dryRun = args.dryRun)
         else:
             logging.info("Empty destination excluded tree")
         logging.info("")
         if tree_unaccounted_destination is not None:
             logging.info("Deleting destination unaccounted tree")
-            fs_destination.removeTree(path_destination, tree_unaccounted_destination, dry_run = args.dry_run)
+            fs_destination.removeTree(path_destination, tree_unaccounted_destination, dryRun = args.dryRun)
         else:
             logging.info("Empty destination unaccounted tree")
         logging.info("")
-    elif args.delete_excluded:
+    elif args.deleteExcluded:
         if tree_excluded_destination is not None:
             logging.info("Deleting destination excluded tree")
-            fs_destination.removeTree(path_destination, tree_excluded_destination, dry_run = args.dry_run)
+            fs_destination.removeTree(path_destination, tree_excluded_destination, dryRun = args.dryRun)
         else:
             logging.info("Empty destination excluded tree")
         logging.info("")
     elif args.delete:
         if tree_unaccounted_destination_non_excluded is not None:
             logging.info("Deleting non-excluded-supporting destination unaccounted tree")
-            fs_destination.removeTree(path_destination, tree_unaccounted_destination_non_excluded, dry_run = args.dry_run)
+            fs_destination.removeTree(path_destination, tree_unaccounted_destination_non_excluded, dryRun = args.dryRun)
         else:
             logging.info("Empty non-excluded-supporting destination unaccounted tree")
         logging.info("")
@@ -867,8 +467,11 @@ if __name__ == "__main__":
             fs_source.normPath,
             fs_destination.joinPaths,
             fs_destination.normPath,
-            dry_run = args.dry_run
+            dryRun = args.dryRun
         )
     else:
         logging.info("Empty copy tree")
     logging.info("")
+
+if __name__ == "__main__":
+    main()
