+++ src/adbsync.py
@@ -1,0 +1,6 @@
+#!/usr/bin/env python3
+
+import ADBSync
+
+if __name__ == "__main__":
+    ADBSync.main()
